Abstract ranklib (#61)
* Merge ranklib.train and ranklib.kcv methods

* Abstracted ranklib and outputs cleared

* adjust solr totq lab for Dan"s comments and remove tqdm import from client

* comment print lines to prevent output overload, now that we use tqdm

* more print cleanup

* rebuild index only when forced if index already exists

* fix bug in index check

* version bump es

* add setup check for py install tools

* to_dataframe method for judgments

* plots and what nots

* fix missing import for ltr.download

* bug fix in blogs

* tests are green

+++ ltr/index.py
@@ -1,19 +1,19 @@
 from ltr.helpers.movies import indexable_movies, noop
 
-def rebuild(client, index, doc_src):
+def rebuild(client, index, doc_src, force = False):
     """ Reload a configuration on disk for each search engine
         (Solr a configset, Elasticsearch a json file)
         and reindex
+    """
 
-        """
-    print("Reconfig from disk...")
-
-    client.delete_index(index)
-    client.create_index(index)
-
-    print("Reindexing...")
-
-    client.index_documents(index,
-                           doc_src=doc_src)
-
-    print('Done')
+    if client.check_index_exists(index):
+        if (force):
+            client.delete_index(index)
+            client.create_index(index)
+            client.index_documents(index, doc_src=doc_src)
+        else:
+            print("Index {} already exists. Use `force = True` to delete and recreate".format(index))
+            return None
+    else:
+        client.create_index(index)
+        client.index_documents(index, doc_src=doc_src)
